[picking_dispatch] directory structure
