+++ picking_dispatch/__openerp__.py
@@ -28,19 +28,21 @@
              'delivery',
              ],
  'website': 'http://www.camptocamp.com/',
- 'data': ['picking_dispatch_view.xml',
-          'picking_dispatch_sequence.xml',
-          'wizard/create_dispatch_view.xml',
-          'wizard/dispatch_assign_picker_view.xml',
-          'wizard/dispatch_start_view.xml',
-          'wizard/check_assign_all_view.xml',
-          'views/report_picking_dispatch.xml',
-          'picking_dispatch_report.xml',
-          'cron_data.xml',
-          'security/ir.model.access.csv',
-          'security/security.xml',
-          # 'picking_dispatch_workflow.xml',
-          ],
+ 'data': [
+     'data/cron.xml',
+     'data/picking_dispatch_sequence.xml',
+     'views/picking_dispatch.xml',
+     'views/product_product.xml',
+     'views/report_picking_dispatch.xml',
+     'views/stock_picking.xml',
+     'views/stock_move.xml',
+     'wizard/create_dispatch_view.xml',
+     'wizard/dispatch_assign_picker_view.xml',
+     'wizard/dispatch_start_view.xml',
+     'wizard/check_assign_all_view.xml',
+     'security/ir.model.access.csv',
+     'security/security.xml',
+ ],
  'tests': [],
  'installable': False,
  'auto_install': False,
