+++ picking_dispatch/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+# @2016 Cyril Gaudin, Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import picking_dispatch
+from . import product_product
+from . import res_company
+from . import stock_move
+from . import stock_picking
