+++ picking_dispatch/models/stock_move.py
@@ -143,44 +143,3 @@ class StockMove(orm.Model):
                 dispatches.add(move.dispatch_id.id)
         dispatch_obj.check_finished(cr, uid, list(dispatches), context)
         return status
-
-
-class StockPicking(orm.Model):
-    _inherit = "stock.picking"
-
-    def _get_related_dispatch(self, cr, uid, ids, field_names, arg=None,
-                              context=None):
-        result = {}
-        if not ids:
-            return result
-        for pick_id in ids:
-            result[pick_id] = []
-        sql = ("SELECT DISTINCT sm.picking_id, sm.dispatch_id "
-               "FROM stock_move sm "
-               "WHERE sm.picking_id in %s AND sm.dispatch_id is NOT NULL")
-        cr.execute(sql, (tuple(ids),))
-        res = cr.fetchall()
-        for picking_id, dispatch_id in res:
-            result[picking_id].append(dispatch_id)
-        return result
-
-    def _search_dispatch_pickings(self, cr, uid, obj, name, args,
-                                  context=None):
-        if not len(args):
-            return []
-        picking_ids = set()
-        for field, symbol, value in args:
-            move_obj = self.pool['stock.move']
-            move_ids = move_obj.search(cr, uid,
-                                       [('dispatch_id', symbol, value)],
-                                       context=context)
-            for move in move_obj.browse(cr, uid, move_ids, context=context):
-                picking_ids.add(move.picking_id.id)
-        return [('id', 'in', list(picking_ids))]
-
-    _columns = {
-        'related_dispatch_ids': fields.function(
-            _get_related_dispatch, fnct_search=_search_dispatch_pickings,
-            type='one2many', relation='picking.dispatch',
-            string='Related Dispatch Picking'),
-    }
