--- picking_dispatch/picking_dispatch_report.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-  <data>
-
-        <report
-            string="Picking dispatch"
-            id="action_report_picking_dispatch"
-            model="picking.dispatch"
-            report_type="qweb-pdf"
-            name="picking_dispatch.report_picking_dispatch"
-            file="picking_dispatch.report_picking_dispatch"
-        />
-
-  </data>
-</openerp>
