--- picking_dispatch/picking_dispatch_workflow.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <!--record id="wkf_picking_dispatch" model="workflow">
-            <field name="name">picking.dispatch.basic</field>
-            <field name="osv">picking.dispatch</field>
-            <field name="on_create">True</field>
-        </record>
-
-        <!- - Activity - ->
-        <record id="act_draft" model="workflow.activity">
-            <field name="wkf_id" ref="wkf_picking_dispatch"/>
-            <field name="flow_start">True</field>
-            <field name="name">draft</field>
-        </record>
-        <record id="act_assigned" model="workflow.activity">
-            <field name="wkf_id" ref="wkf_picking_dispatch"/>
-            <field name="name">assigned</field>
-            <field name="kind">function</field>
-            <field name="action">action_assign</field>
-        </record>
-        <record id="act_progress" model="workflow.activity">
-            <field name="wkf_id" ref="wkf_picking_dispatch"/>
-            <field name="name">progress</field>
-            <field name="kind">function</field>
-            <field name="action">action_progress</field>
-        </record>
-        <record id="act_done" model="workflow.activity">
-            <field name="wkf_id" ref="wkf_picking_dispatch"/>
-            <field name="name">done</field>
-            <field name="flow_stop" eval='True' />
-            <field name="kind">function</field>
-            <field name="action">action_done</field>
-        </record>
-        <record id="act_cancel" model="workflow.activity">
-            <field name="wkf_id" ref="wkf_picking_dispatch"/>
-            <field name="name">cancel</field>
-            <field name="flow_stop" eval='True' />
-            <field name="kind">function</field>
-            <field name="action">action_cancel()</field>
-        </record>
-        <!- - transitions - ->
-        <record id="trans_draft_assigned" model="workflow.transition">
-            <field name="act_from" ref="act_draft"/>
-            <field name="act_to" ref="act_assigned"/>
-            <field name="signal">button_assign</field>
-        </record>
-        <record id="trans_assigned_progress" model="workflow.transition">
-            <field name="act_from" ref="act_assigned"/>
-            <field name="act_to" ref="act_progress"/>
-            <field name="condition">test_start()</field>
-            <field name="signal">button_start</field>
-        </record>
-        <record id="trans_progress_done" model="workflow.transition">
-            <field name="act_from" ref="act_progress"/>
-            <field name="act_to" ref="act_done"/>
-            <field name="condition">test_finished()</field>
-            <field name="signal">button_done</field>
-        </record>
-        <record id="trans_progress_cancel" model="workflow.transition">
-            <field name="act_from" ref="act_progress"/>
-            <field name="act_to" ref="act_cancel"/>
-            <field name="condition">allow_cancel()</field>
-            <field name="signal">button_cancel</field>
-        </record>
-        <record id="trans_draft_cancel" model="workflow.transition">
-            <field name="act_from" ref="act_draft"/>
-            <field name="act_to" ref="act_cancel"/>
-            <field name="condition">allow_cancel()</field>
-            <field name="signal">button_cancel</field>
-        </record>
-        <record id="trans_assigned_cancel" model="workflow.transition">
-            <field name="act_from" ref="act_assigned"/>
-            <field name="act_to" ref="act_cancel"/>
-            <field name="condition">allow_cancel()</field>
-            <field name="signal">button_cancel</field>
-        </record -->
-    </data>
-</openerp>
