+++ picking_dispatch/views/picking_dispatch.xml
@@ -116,92 +116,6 @@
               id='menu_action_picking_dispatch'
               sequence='10'/>
 
-    <!-- Stock picking : add dispatch infos -->
-    <record id="view_picking_form" model="ir.ui.view">
-      <field name="name">stock.picking.form.dispatch</field>
-      <field name="model">stock.picking</field>
-      <field name="inherit_id" ref="stock.view_picking_form"/>
-      <field name="arch" type="xml">
-        <page string="Additional Info" position="after">
-          <page string="Related Dispatch">
-            <field name="related_dispatch_ids" nolabel="1" colspan="4"/>
-          </page>
-        </page>
-      </field>
-    </record>
-
-    <!-- Stock picking : add dispatch search -->
-    <record id="view_picking_internal_search" model="ir.ui.view">
-        <field name="name">stock.picking.search.related_dispatch_ids</field>
-        <field name="model">stock.picking</field>
-        <field name="inherit_id" ref="stock.view_picking_internal_search"/>
-        <field name="arch" type="xml">
-          <field name="name" position="after">
-            <field name="related_dispatch_ids"/>
-          </field>
-      </field>
-    </record>
-
-    <!-- Product View: add description_warehouse -->
-    <record id="product_normal_form_view" model="ir.ui.view">
-      <field name="name">product.normal.form</field>
-      <field name="model">product.product</field>
-      <field name="inherit_id" ref="product.product_normal_form_view"/>
-      <field name="arch" type="xml">
-        <field name="description_purchase"
-               position="after">
-          <separator string="Warehouse Description"/>
-          <field colspan="4" name="description_warehouse" nolabel="1"/>
-        </field>
-      </field>
-    </record>
-
-    <!-- stock.move : display picking.dispatch and allow filtering -->
-    <record id="view_move_tree_receipt_picking" model="ir.ui.view">
-      <field name="name">stock.move.tree2</field>
-      <field name="model">stock.move</field>
-      <field name="inherit_id" ref="stock.view_move_tree_receipt_picking"/>
-      <field name="arch" type="xml">
-        <field name="date" position="after">
-          <field name="dispatch_id" />
-        </field>
-      </field>
-    </record>
-
-    <record id="view_move_form" model="ir.ui.view">
-      <field name="name">stock.move.form</field>
-      <field name="model">stock.move</field>
-      <field name="inherit_id" ref="stock.view_move_form"/>
-      <field name="arch" type="xml">
-        <field name="picking_id" position="after">
-          <field name="dispatch_id"/>
-        </field>
-      </field>
-    </record>
-
-    <record id="view_move_search" model="ir.ui.view">
-      <field name="name">stock.move.search</field>
-      <field name="model">stock.move</field>
-      <field name="inherit_id" ref="stock.view_move_search"/>
-      <field name="arch" type="xml">
-        <field name="partner_id" position="after">
-          <field name="dispatch_id" widget="selection"/>
-        </field>
-      </field>
-    </record>
-
-    <record model="ir.ui.view" id="view_company_form_inherit">
-      <field name="name">res.company.form.inherit</field>
-      <field name="model">res.company</field>
-      <field name="inherit_id" ref="base.view_company_form"/>
-      <field name="arch" type="xml">
-        <notebook position="inside">
-          <page string="Pickings" position="inside">
-            <field name="default_picker_id" colspan ="4"/>
-          </page>
-        </notebook>
-      </field>
-    </record>
 
   </data>
 </openerp>
