+++ picking_dispatch/views/report_picking_dispatch.xml
@@ -1,6 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
+        <report
+            string="Picking dispatch"
+            id="action_report_picking_dispatch"
+            model="picking.dispatch"
+            report_type="qweb-pdf"
+            name="picking_dispatch.report_picking_dispatch"
+            file="picking_dispatch.report_picking_dispatch"
+        />
+
         <template id="report_picking_dispatch">
             <t t-call="report.html_container">
                 <t t-call="report.external_layout">
