+++ picking_dispatch/views/stock_move.xml
@@ -1,0 +1,35 @@
+<odoo>
+    <!-- stock.move : display picking.dispatch and allow filtering -->
+    <record id="view_move_tree_receipt_picking" model="ir.ui.view">
+        <field name="name">stock.move.tree2</field>
+        <field name="model">stock.move</field>
+        <field name="inherit_id" ref="stock.view_move_tree_receipt_picking"/>
+        <field name="arch" type="xml">
+            <field name="date" position="after">
+                <field name="dispatch_id"/>
+            </field>
+        </field>
+    </record>
+
+    <record id="view_move_form" model="ir.ui.view">
+        <field name="name">stock.move.form</field>
+        <field name="model">stock.move</field>
+        <field name="inherit_id" ref="stock.view_move_form"/>
+        <field name="arch" type="xml">
+            <field name="picking_id" position="after">
+                <field name="dispatch_id"/>
+            </field>
+        </field>
+    </record>
+
+    <record id="view_move_search" model="ir.ui.view">
+        <field name="name">stock.move.search</field>
+        <field name="model">stock.move</field>
+        <field name="inherit_id" ref="stock.view_move_search"/>
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="dispatch_id" widget="selection"/>
+            </field>
+        </field>
+    </record>
+</odoo>
