+++ picking_dispatch/views/stock_picking.xml
@@ -1,0 +1,27 @@
+<odoo>
+    <!-- Stock picking : add dispatch infos -->
+    <record id="view_picking_form" model="ir.ui.view">
+        <field name="name">stock.picking.form.dispatch</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_form"/>
+        <field name="arch" type="xml">
+            <page name="extra" position="after">
+                <page string="Related Dispatch">
+                    <field name="related_dispatch_ids" nolabel="1" colspan="4"/>
+                </page>
+            </page>
+        </field>
+    </record>
+
+    <!-- Stock picking : add dispatch search -->
+    <record id="view_picking_internal_search" model="ir.ui.view">
+        <field name="name">stock.picking.search.related_dispatch_ids</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_internal_search"/>
+        <field name="arch" type="xml">
+            <field name="name" position="after">
+                <field name="related_dispatch_ids"/>
+            </field>
+        </field>
+    </record>
+</odoo>
